Add support for Block and fix vardef by creating new envs
In previous versions of the code, I incorrectly pushed new TypeEnvEntries
to gamma in rules IdType and DefTypable, which modified the original gamma.
I think I thought gamma was being passed by value or something, but clearly
it won't be. It's important, then, that we never modify the parameter gamma
in a check function, because it's supposed to be functionally updating
(i.e. returning from a call to check shouldn't have modified my version
of gamma)