Begin adding objects to implementation, but incomplete
The problem I'm having is that constraints placed on an object
effectively need exploding into constraints on the object's members.
I can't do that by inspecting the expression, since the object
might be contained in a variable. In the current implementation,
I explode the constraints while solving a constraint, if I see that
it's a constraint involving two objects