Use "originalObject" to track type modifications
Previously a new type was generated when a property was added to
a type, copying members from the original. Now a type with a
special "originalObj" property pointing to the original object is
used, which allows us to apply constraints to the original type
when appropriate (see test 26). Currently no tests fail, but lots
haven't had results written yet.